[Refactor] 계산기 화면을 한 단계 더 분리
